fix(ui): separate token component to hopefully fix build failure
